requirements.txt
keras==2.14.0
Pillow==10.1.0
scipy==1.11.4
tensorflow==2.14.0
transformers==4.35.2